Refactor preprocessing into preprocessors package
