+++ src/tabpfn/preprocessors/__init__.py
@@ -1,49 +1,90 @@
-from tabpfn.preprocessors.adaptive_quantile_transformer import (
+"""Public preprocessing interface."""
+from __future__ import annotations
+
+from .definitions import (
+    BaseDatasetConfig,
+    ClassifierDatasetConfig,
+    ClassifierEnsembleConfig,
+    EnsembleConfig,
+    PreprocessorConfig,
+    RegressorDatasetConfig,
+    RegressorEnsembleConfig,
+)
+from .presets import (
+    default_classifier_preprocessor_configs,
+    default_regressor_preprocessor_configs,
+    v2_5_classifier_preprocessor_configs,
+    v2_5_regressor_preprocessor_configs,
+    v2_classifier_preprocessor_configs,
+    v2_regressor_preprocessor_configs,
+)
+from .core import (
+    DatasetCollectionWithPreprocessing,
+    balance,
+    build_preprocessing_pipeline,
+    fit_preprocessing,
+    fit_preprocessing_one,
+    generate_classifier_ensemble_configs,
+    generate_index_permutations,
+    generate_regressor_ensemble_configs,
+    get_subsample_indices_for_estimators,
+    transform_labels_one,
+)
+from .steps import (
     AdaptiveQuantileTransformer,
-)
-from tabpfn.preprocessors.add_fingerprint_features_step import (
     AddFingerprintFeaturesStep,
-)
-from tabpfn.preprocessors.differentiable_z_norm_step import DifferentiableZNormStep
-from tabpfn.preprocessors.encode_categorical_features_step import (
+    DifferentiableZNormStep,
     EncodeCategoricalFeaturesStep,
-)
-from tabpfn.preprocessors.kdi_transformer import (
-    KDITransformerWithNaN,
-    get_all_kdi_transformers,
-)
-from tabpfn.preprocessors.nan_handling_polynomial_features_step import (
-    NanHandlingPolynomialFeaturesStep,
-)
-from tabpfn.preprocessors.preprocessing_helpers import (
     FeaturePreprocessingTransformerStep,
-    SequentialFeatureTransformer,
-)
-from tabpfn.preprocessors.remove_constant_features_step import (
+    KDITransformerWithNaN,
+    NanHandlingPolynomialFeaturesStep,
     RemoveConstantFeaturesStep,
-)
-from tabpfn.preprocessors.reshape_feature_distribution_step import (
     ReshapeFeatureDistributionsStep,
+    SafePowerTransformer,
+    SequentialFeatureTransformer,
+    ShuffleFeaturesStep,
+    SquashingScaler,
+    get_all_kdi_transformers,
     get_all_reshape_feature_distribution_preprocessors,
 )
-from tabpfn.preprocessors.safe_power_transformer import SafePowerTransformer
-from tabpfn.preprocessors.shuffle_features_step import ShuffleFeaturesStep
-from tabpfn.preprocessors.squashing_scaler_transformer import SquashingScaler
 
 __all__ = [
     "AdaptiveQuantileTransformer",
     "AddFingerprintFeaturesStep",
+    "BaseDatasetConfig",
+    "ClassifierDatasetConfig",
+    "ClassifierEnsembleConfig",
+    "DatasetCollectionWithPreprocessing",
     "DifferentiableZNormStep",
     "EncodeCategoricalFeaturesStep",
+    "EnsembleConfig",
     "FeaturePreprocessingTransformerStep",
     "KDITransformerWithNaN",
     "NanHandlingPolynomialFeaturesStep",
+    "PreprocessorConfig",
+    "RegressorDatasetConfig",
+    "RegressorEnsembleConfig",
     "RemoveConstantFeaturesStep",
     "ReshapeFeatureDistributionsStep",
     "SafePowerTransformer",
     "SequentialFeatureTransformer",
     "ShuffleFeaturesStep",
     "SquashingScaler",
+    "balance",
+    "build_preprocessing_pipeline",
+    "default_classifier_preprocessor_configs",
+    "default_regressor_preprocessor_configs",
+    "fit_preprocessing",
+    "fit_preprocessing_one",
+    "generate_classifier_ensemble_configs",
+    "generate_index_permutations",
+    "generate_regressor_ensemble_configs",
     "get_all_kdi_transformers",
     "get_all_reshape_feature_distribution_preprocessors",
+    "get_subsample_indices_for_estimators",
+    "transform_labels_one",
+    "v2_5_classifier_preprocessor_configs",
+    "v2_5_regressor_preprocessor_configs",
+    "v2_classifier_preprocessor_configs",
+    "v2_regressor_preprocessor_configs",
 ]
